+++ css/common.css
@@ -1,11 +1,11 @@
 body {
-  background-color: #1b1b1b;
+    background-color: #1b1b1b;
 }
 
 body,
 html {
-  scrollbar-color: #a5a5a5 #1b1b1b;
-  color: #a5a5a5;
+    scrollbar-color: #a5a5a5 #1b1b1b;
+    color: #a5a5a5;
 }
 
 .blind {
@@ -31,12 +31,11 @@ html {
 }
 
 .cards-section .card-poster {
-    width: 240px;
-    height: 360px;
     border-radius: 10px;
-    flex-shrink: 0;
     position: relative;
     z-index: 2;
+    width: calc(20% - 8px);
+    flex-shrink: 0;
 }
 
 .cards-section .card-image-wrap {
@@ -49,7 +48,6 @@ html {
     object-fit: cover;
     width: 100%;
     height: 100%;
-    /* display: block; */
     transition: transform 0.5s ease;
     cursor: pointer;
     border-radius: 10px;
