+++ css/search-result.css
@@ -1,7 +1,7 @@
 #searchResultSection .cards-section {
   display: flex;
   flex-wrap: wrap;
-  gap: 10px;
+  gap: 50px 10px;
 }
 
 #searchResultSection {
@@ -9,36 +9,40 @@
   margin: 80px auto;
 }
 
-/* ----------------------- */
-
-.search-empty-main {
-  width: 1240px;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  justify-content: center;
-  margin: 200px auto 0;
+#searchResultSection .cards-section .card-poster p {
+  padding-top: 8px;
+  padding-right: 12px;
+  table-layout: fixed;
+  font-size: 16px;
+  line-height: 24px;
+  color: #a5a5a5;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  -webkit-line-clamp: 1;
 }
 
-.search-empty-icon {
-  color: white;
-  width: 80px;
-  height: 80px;
-  justify-content: center;
-}
-
-.search-empty-text {
+.no-result-box .search-empty-text {
   margin-top: 12px;
   font-size: 22px;
-  color: white;
+  color: #a5a5a5;
   justify-content: center;
   line-height: 34px;
 }
 
-.error_word {
-  color: #9ab9e8;
+.no-result-box .error-word {
+  color: #3887ff;
 }
 
-.search-empty-text2 {
-  margin-left: 7px;
+.no-result-box {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  padding: 100px 0 120px;
+  text-align: center;
+}
+
+.no-result-box img {
+  width: 74px;
 }
