+++ detail.html
@@ -3,6 +3,7 @@
 <head>
     <meta charset='UTF-8'>
     <title>Energivve(에너자이브)</title>
+    <link rel="shortcut icon" href="./img/favicon.ico" />
     <link rel="stylesheet" href="./css/reset.css"/>
     <link rel="stylesheet" href="./css/common.css"/>
     <link rel="stylesheet" href="./css/detail.css"/>
