+++ index.html
@@ -39,12 +39,10 @@
 
       <section class="cards-section top-rated">
         <div class="cards-section-title">
-          <div class="title">
-            TOP 20
-            <a href="" class="more">
-              더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
-            </a>
-          </div>
+          TOP 20
+          <a href="" class="more">
+            더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
+          </a>
         </div>
         <div
           class="swiper-container second-swiper-container card-carousel">
@@ -60,12 +58,10 @@
 
       <section class="cards-section playing">
         <div class="cards-section-title">
-          <div class="title">
-            NOW PLAYING
-            <a href="" class="more">
-              더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
-            </a>
-          </div>
+          NOW PLAYING
+          <a href="" class="more">
+            더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
+          </a>
         </div>
         <div class="swiper-container card-carousel">
           <ul id="cards-playing" class="swiper-wrapper"></ul>
@@ -80,12 +76,10 @@
 
       <section class="cards-section upcoming">
         <div class="cards-section-title">
-          <div class="title">
-            UPCOMING
-            <a href="" class="more">
-              더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
-            </a>
-          </div>
+          UPCOMING
+          <a href="" class="more">
+            더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
+          </a>
         </div>
         <div class="swiper-container card-carousel">
           <ul id="cards-upcoming" class="swiper-wrapper"></ul>
@@ -99,13 +93,11 @@
       </section>
 
       <section class="cards-section popular">
-        <div class="cards-title">
-          <div class="title">
-            POPULAR
-            <a href="" class="more">
-              더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
-            </a>
-          </div>
+        <div class="cards-section-title">
+          POPULAR
+          <a href="" class="more">
+            더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
+          </a>
         </div>
         <div class="swiper-container card-carousel">
           <ul id="cards-popular" class="swiper-wrapper"></ul>
