+++ search-result.html
@@ -3,6 +3,7 @@
 <head>
     <meta charset="UTF-8" />
     <title>검색 : Energivve(에너자이브)</title>
+    <link rel="shortcut icon" href="./img/favicon.ico" />
     <link rel="stylesheet" href="./css/reset.css" />
     <link rel="stylesheet" href="./css/common.css" />
     <link rel="stylesheet" href="./css/header.css" />
@@ -19,7 +20,8 @@
             href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"
     />
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script defer src="./js/header.js" type="module"></script>
+    <script defer src="./js/common.js" type="module"></script>
+    <script defer src="./js/searchList.js" type="module"></script>
 </head>
 <body>
 <header>
