+++ search.html
@@ -18,7 +18,8 @@
       href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"
     />
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script defer src="./js/header.js" type="module"></script>
+    <script defer src="./js/common.js" type="module"></script>
+    <script defer src="./js/searchList.js" type="module"></script>
   </head>
   <body>
     <header>
