--- searchfailed.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Energivve(에너자이브)</title>
-    <link rel="stylesheet" href="./css/reset.css" />
-    <link rel="stylesheet" href="./css/common.css" />
-    <link rel="stylesheet" href="./css/header.css" />
-    <link rel="stylesheet" href="./css/search.css" />
-    <link rel="stylesheet" href="./css/searchfailed.css" />
-    <link
-      rel="stylesheet"
-      as="style"
-      crossorigin
-      href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard-dynamic-subset.min.css"
-    />
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"
-    />
-    <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-    <script defer src="./js/header.js" type="module"></script>
-  </head>
-  <body>
-    <header>
-      <div class="header-nav">
-        <nav class="header-nav-menu">
-          <ul>
-            <li><a href="#">로그인</a></li>
-            <li><a href="#">이용권</a></li>
-            <li><a href="#">쿠폰·코인</a></li>
-            <li><a href="#">웨이브온</a></li>
-            <li><a href="#">이벤트</a></li>
-          </ul>
-        </nav>
-      </div>
-      <div class="header-section">
-        <a href="./index.html" class="logo">
-          <img src="./img/common/ci-wavve.svg" alt="WAVVE" />
-        </a>
-        <nav class="header-menu">
-          <ul>
-            <li><a href="#" class="on">홈</a></li>
-            <li><a href="#">카테고리</a></li>
-            <li><a href="#">LIVE</a></li>
-            <li><a href="#">MY</a></li>
-          </ul>
-        </nav>
-        <div class="search-container">
-          <img src="img/common/icon-search.svg" class="search-input-icon" />
-          <input
-            placeholder="제목, 장르, 배우로 찾아보세요"
-            type="text"
-            class="search-input-box"
-          />
-          <img src="...." class="search-cancel-icon" />
-        </div>
-      </div>
-    </header>
-    <main class="search-empty-main">
-      <img src="img/common/icon-alert.png" class="search-empty-icon" />
-      <div class="search-empty-text">
-        '
-        <span class="error_word">응애</span>
-        '검색결과가 없어요.
-        <div class="search-empty-text2">이런 검색어는 어떠세요?</div>
-      </div>
-    </main>
-  </body>
-</html>
